Unsub from listener when done, use `useEffect`
This change, first, makes sure to unsub from listening for notifications for the old user when the user changes. This is made possible by `useEffect`'s ability to specify a cleanup function.

`useMemo` should be used when you are computing a value that is needed during a rendering pass.  Typically, the function passed to `useMemo` returns a value, and that value is, in turn, the result of the `useMemo` call each render (or a cached version of it, if the deps didn't change since last render).  Hooks that want to trigger some async action should use `useEffect`.

In general, useMemo usually looks like:

```js
const expensiveValue = useMemo(() => {
  let result;
  // compute result using deps that change infrequently.
  // computation doesn't use async processes, i.e. no api calls,
  // since it needs to return an immediate result.
  return result;
}, [deps, used, to, compute, result]);
```