Notifications! 🔔🔔🔔 (#159)
* Adds notifications.
- Creates new notifications collection for storing each users notis.
- Displays users # of notifications in Badge.
- Shows notifications in Popper
- Notifications exist in header, so state is passed using a context.

* -Restructures notification firestore data
- Restyles notification drop menu, adds in scroll bar
- Adds pagination of notifications

* - Deletes notifications when emojis are taken back
- Adds notifications when comments are added and deletes notifications if comment gets deleted

* - Add responsive styling

* Reworks the notification counter.
- Firestore realtime listeners do NOT work with their count() function. This caused noticeable delays/disconnects between reacting to something and the notification display updating.
- Replaced the running total of notifications with just an indicator of whether you have a notification. Uses direct firestore reads.

* Clean up typos

* Fix alignment of time stamp in the notification drop menu

* Make DeleteNotification wait for deletions

This way, the promise returned by this function won't resolve until all deletions are processed, and by using `Promise.all`, if there happen to be several deletions to do, they will happen in parallel.

* Adjust header spacing

I like having a small margin between the bell and the profile avatar, even on small screens.  This change messes with the spacing a bit to allow this, but still keep things un-cramped in smaller screens.

* Unsub from listener when done, use `useEffect`

This change, first, makes sure to unsub from listening for notifications for the old user when the user changes. This is made possible by `useEffect`'s ability to specify a cleanup function.

`useMemo` should be used when you are computing a value that is needed during a rendering pass.  Typically, the function passed to `useMemo` returns a value, and that value is, in turn, the result of the `useMemo` call each render (or a cached version of it, if the deps didn't change since last render).  Hooks that want to trigger some async action should use `useEffect`.

In general, useMemo usually looks like:

```js
const expensiveValue = useMemo(() => {
  let result;
  // compute result using deps that change infrequently.
  // computation doesn't use async processes, i.e. no api calls,
  // since it needs to return an immediate result.
  return result;
}, [deps, used, to, compute, result]);
```

---------

Co-authored-by: Steve Metz <metzsg@gmail.com>