Update useAnime to use react-query
I didn't earlier, since 99% of the time, this hook wouldn't do any fetching since all DetailedView pages were linked with anime data already in the route state.  Now that notifications uses this hook as well, it seems more common that this will be called, so I've moved it over to react-query.  This should save us redundant