Add info text when reaction buttons are disabled
In this case, I changed the tooltip to let the user know why these buttons are disabled and how to enable them.  Lmk if you have thoughts about copy.

I also sunk the Tooltip into the EmojiReactionChip component since the logic for whether the button is enabled lives in there, so the tooltip text determination should probably be made in the same place.