Add 'Reviewers Often Say' tags on DetailedView
In the spirit of not letting perfect be the enemy of good, this is an intial version of the derived quality tags on anime pages.  I think it is actually pretty cool.  I show up to 4 of the most-commonly described qualities, with a 'see more' button to show the rest.  Future improvements might involve either grouping quality tags by sentiment (positive/negative), including more mixed and negative reviews in the review clustering and labeling to derive more negative tags, and ranking qualities by their uniqueness to make sure interesting ones are surfaced before the 'See more', and generic things like "Strong character development" get moved toward the end of the list.