Fetch random anime via single API call
Since fetching the random items was 6 of the 12 api calls that are made on a /home load, I added support for a 'random' sort via the api's /anime endpoint.  This lets us get a random set of n items with a single api call.  It does return a different set on every call, so I tweaked the react-query settings to get rid of some 'stale' refetches that could happen and make the shelf spontaneously reload without a new page load.