Fix incorrect use of review `docId` for list comment rxns
Based on Ken's feedback, `docId` is already a concatenation of listId and listOwnerId, and so can't be used as argument `listId` in `getRxnTargetForWatchlistComment`.  Since I need to have both of these later, for filling out notifications, I've chosen to just pipe the `listId` through to `Review` from `ReviewContainer` and use that.