Make the 'enter' key submit the email login form
This is just something that has been bugging me.  I like to type my password, then hit enter, instead of clicking a button.  To make this work, I've wrapped the email login form in a `<form>` tag with an `onSubmit` function, and switched the email login button to `type="submit"`.

I also changed the input type for the email input to `type="email"`, which I think will show a special keyboard with an @ sign on mobile, etc.