Fix bug with profile cloning when Firebase User is lost
Problem description and demo in this video: https://drive.google.com/file/d/1_hbLZgrXuul9i9G0NdV7wfWAGLrxWm1j/view?usp=sharing

Basically, Safari can los the LocalUser, which triggers code in `RoutingHelper` to register anonymously.  Unfortunately, the old user's LocalUser data is still present, and the registration code saves that, since usually, during registration, the user does have LocalUser data from onboarding that should be saved.

I've added a solution that takes the user to /login when this happens, since they were already logged in, and lost their authentication.

I've also tried to centralize the 'default LocalUser' definition to one spot, `LocalUserContext`.  Note that there are slight different variations of the default used in various places... `DropMenu` sets avatar and bio to `null` and name to `[]`, while `LocalUserContext` uses `""` for avatar and bio, and doesn't initialize a `name` field.  The logout code in `HandleDialog` seems to have a typo where the field handle is not reset, but instead there is `null: null`.  I've chosen in my implementation to use empty strings, except for handle, and to initialize the name field to empty string there also.  Let's discuss this, though -- not sure if you foresee any issues with that.  There is one catch to setting the name field in the default, which is that it would stomp any name field loaded from firestore in `PopulateFromFirestore` in the `if(!data?.likes)` branch -- I've solved this by expanding the default, and _then_ expanding the data loaded from firestore.  So defaults will all be set, but any values in firestore will be preserved.