Don't refetch recs on window focus.
When the EdwardML tab is selected, if five minutes have passed since page load, react-query will re-fetch the recommendations.  This was pretty much hidden before when the recommendations were deterministic, but now that I am improving that system to involve some randomness, it means sometimes you come back to the window and your recs shift around after a second.  It would be better to make the user explicitly request the refetch, by re-navigating to home, etc.