Fix weird pop-in happening on profile page loads
Happened because useAnimeList would initially get null for its list of ids, and therefore return null.  Once localUser loaded and we got a real list of anime ids, because `keepPreviousData` was set to true, it would continue to return null for the value, and without `isLoading` set to true.  This caused the page to think it was no longer loading and there was no anime data for anything.

We need the `keepPreviousData: true` in `useAnimeList` to handle cases where the user removes items from a watchlist while on a watchlist page without having the whole page revert to a loading state for a second.

A more long-term fix might involve some kind of anime list query cache that doesn't need to make a whole new huge request whenever a profile's anime ids changes by one.  Getting all the profile's anime in one big batch is great during first load, but bad at incremental changes.